refactor server, main function, add request types
